data/requirements.txt
# Google Sheets API
google-auth>=2.0.0
google-api-python-client>=2.0.0

# Data processing
pandas>=2.0.0

# Configuration
pyyaml>=6.0.0

# Date/time utilities
python-dateutil>=2.8.0
